Create guests mvp
    - initial pop_res.py functionality forklifted to /views/admin.py
