+++ backend/backend/urls.py
@@ -19,9 +19,10 @@ from django.urls import path, re_path
 from reservations.views import guests
 from reservations.views import rooms
 from reservations.views import login
+from reservations.views import admin
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
+    #path('admin/', admin.site.urls),
     re_path(r'^api/guests/$', guests.guest_list),
     re_path(r'^api/guests/([0-9])$', guests.guest_detail),
     re_path(r'^api/rooms/$', rooms.room_list),
@@ -32,6 +33,7 @@ urlpatterns = [
     re_path(r'^api/swap_it_up/$', rooms.swap_it_up),
     re_path(r'^api/swap_request/$', rooms.swap_request),
     re_path(r'^api/login_reset/$', login.login_reset),
+    re_path(r'^api/create_guests/$', admin.create_guests),
 ]
