+++ samples/exampleMainAdminList.csv
@@ -1,0 +1,5 @@
+name,email,is_admin,
+Alice,alice@noop.com,False
+Bob,bob@noop.com,False
+Charlie,charlie@noop.com,True
+Index,tyler32bit@gmail.com,True
